+++ dcrjson/walletsvrresults.go
@@ -178,5 +178,5 @@ type ValidateAddressWalletResult struct {
 // GetBestBlockResult models the data from the getbestblock command.
 type GetBestBlockResult struct {
 	Hash   string `json:"hash"`
-	Height int32  `json:"height"`
+	Height int64  `json:"height"`
 }
